+++ docs/src/main/scala/slinky/docs/Navbar.scala
@@ -99,7 +99,7 @@ import scala.scalajs.js
             href := "https://github.com/shadaj/slinky/blob/master/CHANGELOG.md",
             style := smallLinkStyle
           )(
-            "v0.5.1"
+            "v0.5.2"
           ),
           a(
             href := "https://github.com/shadaj/slinky",