+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 + Slinky support for Scala 2.13 is now built with **2.13.0-M5** [PR #202](https://github.com/shadaj/slinky/pull/202)
 + Add support for `getDerivedStateFromProps` from React 16.6 [PR #206](https://github.com/shadaj/slinky/pull/206)
 + Add support for the `Suspense` component from React 16.6 [PR #216](https://github.com/shadaj/slinky/pull/216)
++ Use default parameters to fill in missing props on exported components [PR #221](https://github.com/shadaj/slinky/pull/221)
 
 ### Bug Fixes
 + Fix compile errors when overriding the `Props` or `State` type [PR #220](https://github.com/shadaj/slinky/pull/219)